+++ pinot-core/src/main/java/org/apache/pinot/core/common/NullBitmapUtils.java
@@ -30,13 +30,15 @@ public final class NullBitmapUtils {
   public static void setNullRowIds(RoaringBitmap nullBitmap, ByteArrayOutputStream fixedSizeByteArrayOutputStream,
       ByteArrayOutputStream variableSizeDataByteArrayOutputStream)
       throws IOException {
-    writeInt(fixedSizeByteArrayOutputStream, variableSizeDataByteArrayOutputStream.size());
-    if (nullBitmap.isEmpty()) {
-      writeInt(fixedSizeByteArrayOutputStream, 0);
-    } else {
-      byte[] nullBitmapBytes = ObjectSerDeUtils.ROARING_BITMAP_SER_DE.serialize(nullBitmap);
-      writeInt(fixedSizeByteArrayOutputStream, nullBitmapBytes.length);
-      variableSizeDataByteArrayOutputStream.write(nullBitmapBytes);
+    if (nullBitmap != null) {
+      writeInt(fixedSizeByteArrayOutputStream, variableSizeDataByteArrayOutputStream.size());
+      if (nullBitmap.isEmpty()) {
+        writeInt(fixedSizeByteArrayOutputStream, 0);
+      } else {
+        byte[] nullBitmapBytes = ObjectSerDeUtils.ROARING_BITMAP_SER_DE.serialize(nullBitmap);
+        writeInt(fixedSizeByteArrayOutputStream, nullBitmapBytes.length);
+        variableSizeDataByteArrayOutputStream.write(nullBitmapBytes);
+      }
     }
   }
