+++ pinot-core/src/main/java/org/apache/pinot/core/common/datablock/RowDataBlock.java
@@ -22,6 +22,8 @@ import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.Map;
 import org.apache.pinot.common.utils.DataSchema;
+import org.apache.pinot.core.common.ObjectSerDeUtils;
+import org.roaringbitmap.RoaringBitmap;
 
 
 /**
@@ -46,6 +48,29 @@ public class RowDataBlock extends BaseDataBlock {
       throws IOException {
     super(byteBuffer);
     computeBlockObjectConstants();
+  }
+
+  @Override
+  public RoaringBitmap getNullRowIds(int colId) {
+    // _fixedSizeData stores two ints per col's null bitmap: offset, and length.
+    int position = _numRows * _rowSizeInBytes + colId * Integer.BYTES * 2;
+    if (position >= _fixedSizeData.limit()) {
+      return null;
+    }
+
+    _fixedSizeData.position(position);
+    int offset = _fixedSizeData.getInt();
+    int bytesLength = _fixedSizeData.getInt();
+    RoaringBitmap nullBitmap;
+    if (bytesLength > 0) {
+      _variableSizeData.position(offset);
+      byte[] nullBitmapBytes = new byte[bytesLength];
+      _variableSizeData.get(nullBitmapBytes);
+      nullBitmap = ObjectSerDeUtils.ROARING_BITMAP_SER_DE.deserialize(nullBitmapBytes);
+    } else {
+      nullBitmap = new RoaringBitmap();
+    }
+    return nullBitmap;
   }
 
   protected void computeBlockObjectConstants() {
