+++ pinot-core/src/main/java/org/apache/pinot/core/common/datatable/DataTableImplV4.java
@@ -23,10 +23,8 @@ import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.Map;
 import org.apache.pinot.common.utils.DataSchema;
-import org.apache.pinot.core.common.ObjectSerDeUtils;
 import org.apache.pinot.core.common.datablock.RowDataBlock;
 import org.apache.pinot.spi.annotations.InterfaceStability;
-import org.roaringbitmap.RoaringBitmap;
 
 
 /**
@@ -47,25 +45,6 @@ public class DataTableImplV4 extends RowDataBlock {
   public DataTableImplV4(int numRows, DataSchema dataSchema, Map<String, Map<Integer, String>> dictionaryMap,
       byte[] fixedSizeDataBytes, byte[] variableSizeDataBytes) {
     super(numRows, dataSchema, dictionaryMap, fixedSizeDataBytes, variableSizeDataBytes);
-  }
-
-  @Override
-  public RoaringBitmap getNullRowIds(int colId) {
-    // _fixedSizeData stores two ints per col's null bitmap: offset, and length.
-    int position = _numRows * _rowSizeInBytes + colId * Integer.BYTES * 2;
-    _fixedSizeData.position(position);
-    int offset = _fixedSizeData.getInt();
-    int bytesLength = _fixedSizeData.getInt();
-    RoaringBitmap nullBitmap;
-    if (bytesLength > 0) {
-      _variableSizeData.position(offset);
-      byte[] nullBitmapBytes = new byte[bytesLength];
-      _variableSizeData.get(nullBitmapBytes);
-      nullBitmap = ObjectSerDeUtils.ROARING_BITMAP_SER_DE.deserialize(nullBitmapBytes);
-    } else {
-      nullBitmap = new RoaringBitmap();
-    }
-    return nullBitmap;
   }
 
   @Override
